+++ src/commands/restartSorbet.ts
@@ -1,20 +1,19 @@
 import * as vscode from 'vscode';
 import { safeActiveTextEditorUri } from '../common/utils';
 import { SorbetExtensionContext } from '../sorbetExtensionContext';
-import { LspStatus } from '../types';
-import { anySorbetWorkspace } from '../workspaceUtils';
 import { OPEN_SETTINGS_ID } from './commandIds';
 import { getTargetWorkspaceUri } from './utils';
 
 export async function restartSorbet(
   context: SorbetExtensionContext,
+  action: 'start' | 'stop' | 'restart',
   pathOrUri?: string | vscode.Uri,
 ) {
   const uri = await getTargetWorkspaceUri(pathOrUri ?? safeActiveTextEditorUri(), {
     forceSorbetWorkspace: true,
   });
   if (!uri) {
-    context.log.debug('RestartSorbet: No Sorbet workspace.');
+    context.log.debug('No target workspace.');
     return; // No target workspace
   }
 
@@ -23,18 +22,28 @@ export async function restartSorbet(
     context.log.info('RestartSorbet: No Sorbet client for selected workspace.', uri.toString(true));
     return;
   }
-  if (client.status === LspStatus.Disabled && !client.isEnabledByConfiguration()) {
-    await showDisabledConfigurationNotification(uri);
-    return;
-  } else if (!vscode.workspace.workspaceFolders?.length) {
-    await showNoWorkspaceNotification();
-    return;
-  } else if (!(await anySorbetWorkspace())) {
-    await showMissingSorbetWorkspaceNotification();
-    return;
-  }
 
-  await client.restart();
+  switch (action) {
+    case 'restart':
+      if (!client.isEnabledByConfiguration()) {
+        await showDisabledConfigurationNotification(uri);
+        return;
+      }
+      await client.restart();
+      break;
+    case 'start':
+      if (!client.isEnabledByConfiguration()) {
+        await showDisabledConfigurationNotification(uri);
+        return;
+      }
+      await client.start();
+      break;
+    case 'stop':
+      await client.stop();
+      break;
+    default:
+      throw new Error(`Unknown action: ${action}`);
+  }
 }
 
 async function showDisabledConfigurationNotification(uri: vscode.Uri) {
@@ -46,19 +55,4 @@ async function showDisabledConfigurationNotification(uri: vscode.Uri) {
   if (selection === updateConfigItem) {
     await vscode.commands.executeCommand(OPEN_SETTINGS_ID, uri, 'sorbetto.sorbetLspConfiguration');
   }
-}
-
-async function showMissingSorbetWorkspaceNotification() {
-  const setupWorkspaceItem: vscode.MessageItem = { title: 'Setup' };
-  const selection = await vscode.window.showWarningMessage(
-    'Workspace is not setup to run Sorbet.',
-    setupWorkspaceItem,
-  );
-  if (selection === setupWorkspaceItem) {
-    await vscode.commands.executeCommand('sorbetto.setup.workspace');
-  }
-}
-
-async function showNoWorkspaceNotification() {
-  await vscode.window.showWarningMessage('No workspace is open, Sorbet cannot be started');
 }