+++ src/sorbetLanguageStatus.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode';
 import { OPEN_SETTINGS_ID, SHOW_OUTPUT_ID, SORBET_RESTART_ID } from './commands/commandIds';
-import { debounce, onSafeActiveTextEditorChanged, safeActiveTextEditorUri } from './common/utils';
+import { onSafeActiveTextEditorChanged, safeActiveTextEditorUri } from './common/utils';
 import { SORBET_DOCUMENT_SELECTOR } from './lsp/documentSelectors';
 import { LspConfigurationType } from './lspClient/configuration/lspConfigurationType';
 import { SorbetClient } from './lspClient/sorbetClient';
@@ -49,9 +49,7 @@ export class SorbetLanguageStatus implements vscode.Disposable {
     this.setStatus({ status: 'Disabled', command: StartCommand });
 
     this.disposables = [
-      onSafeActiveTextEditorChanged(
-        debounce((editor) => this.handleEditorOrStatusChange(undefined, editor)),
-      ),
+      onSafeActiveTextEditorChanged((editor) => this.handleEditorOrStatusChange(undefined, editor)),
       this.context.statusProvider.onShowOperation(({ client }) =>
         this.handleEditorOrStatusChange(client),
       ),